Fix database issues that where causing characters to not save or be able to be created.

+++ Source/Game/Entities/Player/Player.DB.cs
@@ -1782,7 +1782,7 @@ namespace Game.Entities
 
                 ushort value = skillInfoField.SkillRank[pair.Value.Pos];
                 ushort max = skillInfoField.SkillMaxRank[pair.Value.Pos];
-                byte professionSlot = (byte)GetProfessionSlotFor(pair.Key);
+                int professionSlot = GetProfessionSlotFor(pair.Key);
 
                 switch (pair.Value.State)
                 {
